# notificationService/utils/swagger.yaml
openapi: 3.0.0
info:
  title: Notification System API
  version: 1.0.0
  description: API documentation for the Notification System

servers:
  - url: 'http://localhost:8080'

tags:
  - name: Notifications
    description: Notification related endpoints

paths:
  /notifications:
    post:
      summary: Create a new notification
      tags: [Notifications]
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - userId
                - message
              properties:
                userId:
                  type: string
                  description: ID of the user to receive the notification
                message:
                  type: string
                  description: Notification message content
      responses:
        201:
          description: Notification created successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: string
                  userId:
                    type: string
                  message:
                    type: string
        400:
          description: Bad request
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string

    get:
      summary: Get notifications for the authenticated user
      tags: [Notifications]
      security:
        - bearerAuth: []
      parameters:
        - name: page
          in: query
          description: Page number for pagination
          required: false
          schema:
            type: integer
        - name: limit
          in: query
          description: Number of items per page
          required: false
          schema:
            type: integer
      responses:
        200:
          description: List of notifications
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    id:
                      type: string
                    userId:
                      type: string
                    message:
                      type: string
        400:
          description: Bad request
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string

  /notifications/{id}:
    get:
      summary: Get a specific notification detail by ID
      tags: [Notifications]
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          description: ID of the notification
          required: true
          schema:
            type: string
      responses:
        200:
          description: Notification details
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: string
                  userId:
                    type: string
                  message:
                    type: string
        400:
          description: Bad request
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
        404:
          description: Notification not found
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string

    put:
      summary: Mark a notification as read
      tags: [Notifications]
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          description: ID of the notification
          required: true
          schema:
            type: string
      responses:
        200:
          description: Notification marked as read
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: string
                  userId:
                    type: string
                  message:
                    type: string
                  read:
                    type: boolean
        400:
          description: Bad request
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
        404:
          description: Notification not found
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT